requirements.txt
Django==1.6.5
gunicorn==18.0
pluct==0.4.1
requests==2.3.0
django-gravatar2==1.1.3
python-dateutil==2.2
pygments==1.6
